+++ app/services/http-handler/http-handler.serv.ts
@@ -1,13 +1,14 @@
 ﻿//import {Router} from '@angular/router';
-import {Injectable} from '@angular/core';
-import {Http, Headers, HTTP_PROVIDERS, RequestOptionsArgs, Request, Response, URLSearchParams} from '@angular/http';
-import {Observable} from 'rxjs/Observable';
-import {HelperService} from '../../services/helper/helper.serv';
-import 'rxjs/Rx'; //for map
+import {Injectable} from "@angular/core";
+import {Router} from "@angular/router";
+import {Http, Headers, RequestOptionsArgs, Response, URLSearchParams} from "@angular/http";
+import {Observable} from "rxjs/Observable";
+import {HelperService} from "../../services/helper/helper.serv";
+import "rxjs/Rx"; //for map
 
 @Injectable()
 export class HttpHandlerService {
-    constructor(private http: Http) {
+    constructor(private http: Http, private router: Router) {
         console.log('constructor HttpHandlerService');
         this.serviceBase = HelperService.getServiceBase()
     }
@@ -17,46 +18,50 @@ export class HttpHandlerService {
     //use http get to retrieve an object of type T
     //parameters: an array of name / value pairs
     getObject<T>(parameters: modSharedTypes.IHttpParameter[], url: string, includeToken: boolean): Observable<T> {
-        var options: RequestOptionsArgs = this.getOptions(parameters, includeToken);
-        var obs = this.http.get(this.serviceBase + url, options).map(res => res.json());
-        return obs;
+        if (HelperService.tokenIsValid()) {
+            var options: RequestOptionsArgs = this.getOptions(parameters, includeToken);
+            var obs = this.http.get(this.serviceBase + url, options).map(res => res.json());
+            return obs;
+        } else {
+            HelperService.sendToLogin(this.router)
+        }
     }
 
     deleteObject(parameters: modSharedTypes.IHttpParameter[], url: string): Observable<Response> {
-        //deleteObject<T>(parameters: modSharedTypes.IHttpParameter[], url: string): Observable < any > {
-        var options: RequestOptionsArgs = this.getOptions(parameters, true);
-        return this.http.delete(this.serviceBase + url, options);
+        if (HelperService.tokenIsValid()) {
+            var options: RequestOptionsArgs = this.getOptions(parameters, true);
+            return this.http.delete(this.serviceBase + url, options);
+        } else {
+            HelperService.sendToLogin(this.router)
+        }
     }
 
     //use http post to send an object 
     postObject(parameterObj: Object, url: string, includeToken: boolean = true): Observable<Response> {
-        //postObject<T>(parameterObj: Object, url: string, includeToken: boolean = true): Observable < Response > {
-        var options: RequestOptionsArgs = this.postOptions(includeToken);
-        var s = JSON.stringify(parameterObj);
-        return this.http.post(this.serviceBase + url, s, options);
+        if (HelperService.tokenIsValid()) {
+            var options: RequestOptionsArgs = this.postOptions(includeToken);
+            var s = JSON.stringify(parameterObj);
+            return this.http.post(this.serviceBase + url, s, options);
+        } else {
+            HelperService.sendToLogin(this.router)
+        }
     }
     //use http put to send an object 
     putObject(parameterObj: Object, url: string, includeToken: boolean = true): Observable<Response> {
-        var options: RequestOptionsArgs = this.postOptions(includeToken);
-        var s = JSON.stringify(parameterObj);
-        return this.http.put(this.serviceBase + url, s, options);
+        if (HelperService.tokenIsValid()) {
+            var options: RequestOptionsArgs = this.postOptions(includeToken);
+            var s = JSON.stringify(parameterObj);
+            return this.http.put(this.serviceBase + url, s, options);
+        } else {
+            HelperService.sendToLogin(this.router)
+        }
     }
 
-
-    private logError() {
-        console.log('get Entities failed');
-
-    };
-
-    private parseResponse(res: Response) {
-        return res.json();
-    }
-
-    private getHeaders(includeToken: boolean): Headers {
+    getHeaders(includeToken: boolean): Headers {
         var headers = new Headers();
-        //get login token from storage and add headers
         var token: string;
 
+        //get login token from storage and add headers
         headers.append('Accept', 'application/json');
         headers.append('Content-Type', 'application/json');
 
@@ -68,7 +73,7 @@ export class HttpHandlerService {
     }
 
 
-    private getParameters(parameters: modSharedTypes.IHttpParameter[]) {
+    getParameters(parameters: modSharedTypes.IHttpParameter[]) {
         var params: URLSearchParams = new URLSearchParams(), i: number;
         for (i = 0; i < parameters.length; i++) {
             params.append(parameters[i].name, parameters[i].value);
@@ -83,6 +88,7 @@ export class HttpHandlerService {
         var params: URLSearchParams = this.getParameters(parameters);
         options.search = params;
         options.headers = headers;
+        options.body='';
         return options;
     }
